Yet more warning fixes

+++ game/game_libs/cl_dlls/health.cpp
@@ -191,18 +191,18 @@ int CHudHealth::Draw(float flTime)
 	// Has health changed? Flash the health #
 	if ( m_fFade )
 	{
-		m_fFade -= (gHUD.m_flTimeDelta * 20);
+		m_fFade -= static_cast<float>(gHUD.m_flTimeDelta * 20);
 		if ( m_fFade <= 0 )
 		{
-			a = MIN_ALPHA;
+			a = static_cast<int>(MIN_ALPHA);
 			m_fFade = 0;
 		}
 
 		// Fade the health number back to dim
-		a = MIN_ALPHA + (m_fFade / FADE_TIME) * 128;
+		a = static_cast<int>(MIN_ALPHA + (m_fFade / FADE_TIME) * 128);
 	}
 	else
-		a = MIN_ALPHA;
+		a = static_cast<int>(MIN_ALPHA);
 
 	// If health is getting low, make it bright red
 	if ( m_iHealth <= 15 )
@@ -277,7 +277,7 @@ void CHudHealth::CalcDamageDirection(vec3_t vecFrom)
 		}
 		else
 		{
-			float f = fabs(side);
+			float f = fabsf(side);
 			if ( f > 0.3 )
 				m_fAttackRear = Max(m_fAttackRear, f);
 		}
@@ -289,7 +289,7 @@ void CHudHealth::CalcDamageDirection(vec3_t vecFrom)
 		}
 		else
 		{
-			float f = fabs(front);
+			float f = fabsf(front);
 			if ( f > 0.3 )
 				m_fAttackLeft = Max(m_fAttackLeft, f);
 		}
@@ -307,13 +307,13 @@ int CHudHealth::DrawPain(float)
 	// TODO:  get the shift value of the health
 	a = 255;  // max brightness until then
 
-	float fFade = gHUD.m_flTimeDelta * 2;
+	float fFade = static_cast<float>(gHUD.m_flTimeDelta * 2);
 
 	// SPR_Draw top
 	if ( m_fAttackFront > 0.4 )
 	{
 		GetPainColor(r, g, b);
-		shade = a * Max(m_fAttackFront, 0.5f);
+		shade = static_cast<int>(a * Max(m_fAttackFront, 0.5f));
 		ScaleColors(r, g, b, shade);
 		SPR_Set(m_hSprite, r, g, b);
 
@@ -328,7 +328,7 @@ int CHudHealth::DrawPain(float)
 	if ( m_fAttackRight > 0.4 )
 	{
 		GetPainColor(r, g, b);
-		shade = a * Max(m_fAttackRight, 0.5f);
+		shade = static_cast<int>(a * Max(m_fAttackRight, 0.5f));
 		ScaleColors(r, g, b, shade);
 		SPR_Set(m_hSprite, r, g, b);
 
@@ -343,7 +343,7 @@ int CHudHealth::DrawPain(float)
 	if ( m_fAttackRear > 0.4 )
 	{
 		GetPainColor(r, g, b);
-		shade = a * Max(m_fAttackRear, 0.5f);
+		shade = static_cast<int>(a * Max(m_fAttackRear, 0.5f));
 		ScaleColors(r, g, b, shade);
 		SPR_Set(m_hSprite, r, g, b);
 
@@ -358,7 +358,7 @@ int CHudHealth::DrawPain(float)
 	if ( m_fAttackLeft > 0.4 )
 	{
 		GetPainColor(r, g, b);
-		shade = a * Max(m_fAttackLeft, 0.5f);
+		shade = static_cast<int>(a * Max(m_fAttackLeft, 0.5f));
 		ScaleColors(r, g, b, shade);
 		SPR_Set(m_hSprite, r, g, b);
